The issue of Rx buffer overflow resolved.
When Rx buffer overflows, available returns 0. In fact if the head passes ahead of the tail, count will become 0 as the count is always computed as (head - tail). I think this is the best experience right now because I have no control over the head position and I do not get notified when the head goes forward in the buffer.